Add PDB reader/writer code

+++ src/DotNet/Writer/ModuleWriterBase.cs
@@ -25,6 +25,7 @@
 using System.Collections.Generic;
 using System.IO;
 using dnlib.IO;
+using dnlib.DotNet.Pdb;
 using dnlib.PE;
 using dnlib.W32Resources;
 using dnlib.DotNet.MD;
@@ -173,6 +174,33 @@ namespace dnlib.DotNet.Writer {
 		}
 
 		/// <summary>
+		/// Set it to <c>true</c> to enable writing a PDB file. Default is <c>false</c> (a PDB file
+		/// won't be written to disk).
+		/// </summary>
+		public bool WritePdb { get; set; }
+
+		/// <summary>
+		/// PDB file name. If it's <c>null</c> a PDB file with the same name as the output assembly
+		/// will be created but with a PDB extension. <see cref="WritePdb"/> must be <c>true</c> or
+		/// this property is ignored.
+		/// </summary>
+		public string PdbFileName { get; set; }
+
+		/// <summary>
+		/// PDB stream. If this is initialized, then you should also set <see cref="PdbFileName"/>
+		/// to the name of the PDB file since the file name must be written to the PE debug directory.
+		/// <see cref="WritePdb"/> must be <c>true</c> or this property is ignored.
+		/// </summary>
+		public Stream PdbStream { get; set; }
+
+		/// <summary>
+		/// If <see cref="PdbFileName"/> or <see cref="PdbStream"/> aren't enough, this can be used
+		/// to create a new <see cref="ISymbolWriter2"/> instance. <see cref="WritePdb"/> must be
+		/// <c>true</c> or this property is ignored.
+		/// </summary>
+		public CreatePdbSymbolWriterDelegate CreatePdbSymbolWriter { get; set; }
+
+		/// <summary>
 		/// Default constructor
 		/// </summary>
 		protected ModuleWriterOptionsBase() {
@@ -306,6 +334,13 @@ namespace dnlib.DotNet.Writer {
 	}
 
 	/// <summary>
+	/// Creates a new <see cref="ISymbolWriter2"/> instance
+	/// </summary>
+	/// <param name="writer">Module writer</param>
+	/// <returns>A new <see cref="ISymbolWriter2"/> instance</returns>
+	public delegate ISymbolWriter2 CreatePdbSymbolWriterDelegate(ModuleWriterBase writer);
+
+	/// <summary>
 	/// Module writer base class
 	/// </summary>
 	public abstract class ModuleWriterBase : IMetaDataListener, ILogger {
@@ -323,6 +358,8 @@ namespace dnlib.DotNet.Writer {
 		protected const uint DEFAULT_STRONGNAMESIG_ALIGNMENT = 16;
 		/// <summary>Default COR20 header alignment</summary>
 		protected const uint DEFAULT_COR20HEADER_ALIGNMENT = 4;
+		/// <summary>Default debug directory alignment</summary>
+		protected const uint DEFAULT_DEBUGDIRECTORY_ALIGNMENT = 4;
 
 		/// <summary>See <see cref="DestinationStream"/></summary>
 		protected Stream destStream;
@@ -339,6 +376,10 @@ namespace dnlib.DotNet.Writer {
 		/// <summary>Offset where the module is written. Usually 0.</summary>
 		protected long destStreamBaseOffset;
 		IModuleWriterListener listener;
+		/// <summary>Debug directory</summary>
+		protected DebugDirectory debugDirectory;
+
+		string createdPdbFileName;
 
 		/// <summary>
 		/// Strong name signature
@@ -423,6 +464,13 @@ namespace dnlib.DotNet.Writer {
 		public abstract PESection RsrcSection { get; }
 
 		/// <summary>
+		/// Gets the debug directory or <c>null</c> if there's none
+		/// </summary>
+		public DebugDirectory DebugDirectory {
+			get { return debugDirectory; }
+		}
+
+		/// <summary>
 		/// <c>true</c> if <c>this</c> is a <see cref="NativeModuleWriter"/>, <c>false</c> if
 		/// <c>this</c> is a <see cref="ModuleWriter"/>.
 		/// </summary>
@@ -450,6 +498,8 @@ namespace dnlib.DotNet.Writer {
 		}
 
 		static void DeleteFileNoThrow(string fileName) {
+			if (string.IsNullOrEmpty(fileName))
+				return;
 			try {
 				File.Delete(fileName);
 			}
@@ -575,6 +625,99 @@ namespace dnlib.DotNet.Writer {
 		protected void StrongNameSign(long snSigOffset) {
 			var snSigner = new StrongNameSigner(destStream, destStreamBaseOffset);
 			snSigner.WriteSignature(TheOptions.StrongNameKey, snSigOffset);
+		}
+
+		/// <summary>
+		/// Write the PDB file. The caller should send the PDB events before and after calling this
+		/// method.
+		/// </summary>
+		protected void WritePdbFile() {
+			if (!TheOptions.WritePdb)
+				return;
+			if (debugDirectory == null)
+				throw new InvalidOperationException("debugDirectory is null but WritePdb is true");
+
+			var pdbState = Module.PdbState;
+			if (pdbState == null) {
+				Error("TheOptions.WritePdb is true but module has no PdbState");
+				debugDirectory.DontWriteAnything = true;
+				return;
+			}
+
+			var symWriter = GetSymbolWriter2();
+			if (symWriter == null) {
+				Error("Could not create a PDB symbol writer. A Windows OS might be required.");
+				debugDirectory.DontWriteAnything = true;
+				return;
+			}
+
+			var pdbWriter = new PdbWriter(symWriter, pdbState, metaData);
+			try {
+				pdbWriter.Logger = TheOptions.Logger;
+				pdbWriter.Write();
+
+				debugDirectory.Data = pdbWriter.GetDebugInfo(out debugDirectory.debugDirData);
+				debugDirectory.TimeDateStamp = GetTimeDateStamp();
+				pdbWriter.Dispose();
+			}
+			catch {
+				pdbWriter.Dispose();
+				DeleteFileNoThrow(createdPdbFileName);
+				throw;
+			}
+		}
+
+		uint GetTimeDateStamp() {
+			var td = TheOptions.PEHeadersOptions.TimeDateStamp;
+			if (td.HasValue)
+				return (uint)td;
+			TheOptions.PEHeadersOptions.TimeDateStamp = PEHeadersOptions.CreateNewTimeDateStamp();
+			return (uint)TheOptions.PEHeadersOptions.TimeDateStamp;
+		}
+
+		ISymbolWriter2 GetSymbolWriter2() {
+			if (TheOptions.CreatePdbSymbolWriter != null) {
+				var writer = TheOptions.CreatePdbSymbolWriter(this);
+				if (writer != null)
+					return writer;
+			}
+
+			if (TheOptions.PdbStream != null) {
+				return SymbolWriterCreator.Create(TheOptions.PdbStream,
+							TheOptions.PdbFileName ??
+							GetStreamName(TheOptions.PdbStream) ??
+							GetDefaultPdbFileName());
+			}
+
+			if (!string.IsNullOrEmpty(TheOptions.PdbFileName)) {
+				createdPdbFileName = TheOptions.PdbFileName;
+				return SymbolWriterCreator.Create(createdPdbFileName);
+			}
+
+			createdPdbFileName = GetDefaultPdbFileName();
+			return SymbolWriterCreator.Create(createdPdbFileName);
+		}
+
+		static string GetStreamName(Stream stream) {
+			var fs = stream as FileStream;
+			return fs == null ? null : fs.Name;
+		}
+
+		string GetDefaultPdbFileName() {
+			var destFileName = GetDestinationFileName();
+			if (string.IsNullOrEmpty(destFileName)) {
+				Error("TheOptions.WritePdb is true but it's not possible to guess the default PDB file name. Set PdbFileName to the name of the PDB file.");
+				return null;
+			}
+
+			return Path.ChangeExtension(destFileName, "pdb");
+		}
+
+		string GetDestinationFileName() {
+			var fs = destStream as FileStream;
+			if (fs == null)
+				return null;
+			return fs.Name;
 		}
 
 		/// <inheritdoc/>
